+++ src/components/fields/FileField.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 import { Input } from "@/components/common/Input";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/common/Button";
 import { cn } from "@/lib/utils";
 
 export type FileFieldProps = {
